requirements.txt
asgiref==3.5.2
cloudinary==1.29.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.13
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-summernote==0.8.20.0
gunicorn==20.1.0
lxml==4.8.0
multitasking==0.0.10
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
polygon-api-client==1.0.2
psycopg2==2.9.3
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
websockets==10.3
yfinance==0.1.70
